Isolate all unpacking related functions to their own source file in preparation for GUI implementation
