+++ CMakeLists.txt
@@ -1,10 +1,11 @@
-cmake_minimum_required(VERSION 3.21)
+cmake_minimum_required(VERSION 3.22)
 project(sp64-tt LANGUAGES CXX)
 
 add_subdirectory(external)
-add_executable(sp64-tt src/main.cpp)
+add_executable(sp64-tt src/main.cpp
+                       src/Unpacker.cpp)
 target_link_libraries(sp64-tt PRIVATE spdlog SFML::Graphics)
-target_compile_features(sp64-tt PRIVATE cxx_std_17)
+target_compile_features(sp64-tt PRIVATE cxx_std_20)
 
 if(CMAKE_CXX_COMPILER_ID MATCHES "(GNU|Clang)")
     target_compile_options(sp64-tt PRIVATE -Werror -Wall -Wextra -Wpedantic -Wshadow -Wconversion -Wsign-conversion)
