+++ src/Unpacker.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <filesystem>
+
+namespace tt {
+[[nodiscard]] bool UnpackTextures(const std::filesystem::path& packedAssetPath);
+}
